### 2025-04-02 #7979 @editorialbot
[PRE REVIEW]: CaNS-Fizzy: A GPU-accelerated finite difference solver for turbulent two-phase flows
**Submitting author:** @gianlupo (<a href="http://orcid.org/0000-0002-1095-118X">Giandomenico Lupo</a>)
**Repository:** https://github.com/CaNS-World/CaNS-Fizzy
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0
