### 2025-04-18 #8076 @piyueh
As a reviewer, the first step is to create a checklist for your review by entering `@editorialbot generate my checklist` as a new comment in this thread.

These checklists contain the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. The first comment in this thread also contains links to the JOSS reviewer guidelines. Or you can find them here:
* https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
* https://joss.readthedocs.io/en/latest/review_criteria.html
* https://joss.readthedocs.io/en/latest/review_checklist.html

We aim for reviews to be completed within about 4-6 weeks. Please let me know if any of you require some more time. We can also use `@editorialbot` to set automatic reminders if you know you'll be away for a known period of time.
