### 2025-04-18 #8076 @archermarx
Hey all, just completed my read through the paper and repository. I'll work on running the code this afternoon.

A few comments to start:
1. It seems like the automated tests may not be working. Looking at the workflow runs [here,](https://github.com/CaNS-World/CaNS-Fizzy/actions) it seems that something may be misconfigured with github actions. It would be good to get those running and see that green checkmark before acceptance.
2. The paper does not mention or compare to other similar packages (State of the Field). A sentence or two mentioning what else is out there and how your code is different from them would be good. I gather from context that your GPU parallelization and Poisson solve approach may be the main differentiators.
3. Found a typo: Line 19: "as base" -> "as a base"
4. Does the code only work for NVIDIA GPUs? If so it would be good to make that clear, or if not, highlight that.

As part of my reviewer checklist, I am supposed to verify the performance of the application. Lacking free access to a large supercomputer, I can't verify the weak scaling performance very easily. @piyueh, what would constitute verifying performance claims in this case?
