### 2025-04-18 #8076 @piyueh
> ... Lacking free access to a large supercomputer, I can't verify the weak scaling performance very easily. ...

@archermarx thanks for raising this issue.

@gianlupo Could you highlight some smaller scaling tests or performance benchmarks that reviewers can run locally to verify a subset of the performance claims? It's not necessary for reviewers (@archermarx @akashdhruv) to reproduce **_all_** the results/claims, but it's important to give them a way to build confidence in the claims that they can't directly verify due to hardware limitations. Some locally feasible scaling tests or lightweight benchmarks would be really helpful.
