### 2025-04-24 d465560
Fixed typo in line 19 of paper.md
