+++ paper/paper.md
@@ -29,7 +29,7 @@ bibliography: paper.bib
 
 # Summary
 
-CaNS-Fizzy -- Fizzy for short -- is a GPU-accelerated numerical solver for massively-parallel Direct Numerical Simulations (DNS) of incompressible two-phase flows. A DNS enables direct access to all flow quantities, resolved in time and space at all relevant continuum scales. The resulting numerical experiments provide complete data sets for the analysis of the detailed mechanisms underlying the flow, particularly the interaction between the chaotic and multi-scale dynamics of turbulence and the interface movement and deformation. The insights gained can guide the design and operation of various applications, such as boiling heat transfer, liquid-liquid extraction, gas-liquid reactors, absorption and stripping columns, distillation columns, liquid combustion appliances, in all of which the rate of heat and mass transfer between phases is proportional to the interfacial area. Fizzy's two-phase capabilities were implemented using the efficient, GPU-accelerated Navier-Stokes solver CaNS [@Costa:2018] as base.
+CaNS-Fizzy -- Fizzy for short -- is a GPU-accelerated numerical solver for massively-parallel Direct Numerical Simulations (DNS) of incompressible two-phase flows. A DNS enables direct access to all flow quantities, resolved in time and space at all relevant continuum scales. The resulting numerical experiments provide complete data sets for the analysis of the detailed mechanisms underlying the flow, particularly the interaction between the chaotic and multi-scale dynamics of turbulence and the interface movement and deformation. The insights gained can guide the design and operation of various applications, such as boiling heat transfer, liquid-liquid extraction, gas-liquid reactors, absorption and stripping columns, distillation columns, liquid combustion appliances, in all of which the rate of heat and mass transfer between phases is proportional to the interfacial area. Fizzy's two-phase capabilities were implemented using the efficient, GPU-accelerated Navier-Stokes solver CaNS [@Costa:2018] as a base.
 
 # Statement of need
 
