+++ paper/paper.bib
@@ -86,3 +86,53 @@ year = {2023},
 doi = {10.21105/joss.05813},
 author = {S. Rolfo and C. Flageul and P. Bartholomew and F. Spiga and S. Laizet}
 }
+
+@article{gerris,
+title = {An accurate adaptive solver for surface-tension-driven interfacial flows},
+journal = {Journal of Computational Physics},
+volume = {228},
+number = {16},
+pages = {5838-5866},
+year = {2009},
+doi = {10.1016/j.jcp.2009.04.042},
+author = {Popinet, S.}
+}
+
+@article{basilisk,
+title = {A quadtree-adaptive multigrid solver for the Serre–Green–Naghdi equations},
+journal = {Journal of Computational Physics},
+volume = {302},
+pages = {336-358},
+year = {2015},
+doi = {10.1016/j.jcp.2015.09.009},
+author = {Popinet, S.}
+}
+
+@article{boilingfoam,
+title = {Computational study of bubble, thin-film dynamics and heat transfer during flow boiling in non-circular microchannels},
+journal = {Applied Thermal Engineering},
+volume = {238},
+pages = {122039},
+year = {2024},
+doi = {10.1016/j.applthermaleng.2023.122039},
+author = {F. Municchi and C.N. Markides and O.K. Matar and M. Magnini}
+}
+
+@article{flow36,
+title = {FLOW36: A spectral solver for phase-field based multiphase turbulence simulations on heterogeneous computing architectures},
+journal = {Computer Physics Communications},
+pages = {109640},
+year = {2025},
+doi = {10.1016/j.cpc.2025.109640},
+author = {Roccon, A. and Soligo, G. and Soldati, A.}
+}
+
+@article{thinc,
+title = {Toward efficient and accurate interface capturing on arbitrary hybrid unstructured grids: The THINC method with quadratic surface representation and Gaussian quadrature},
+journal = {Journal of Computational Physics},
+volume = {349},
+pages = {415-440},
+year = {2017},
+doi = {10.1016/j.jcp.2017.08.028},
+author = {Xie, B. and Xiao, F.}
+}
