### 2025-06-10 #8076 @archermarx
I've finished reviewing the paper and will download and install the software this afternoon. If all goes well I'll be ready to accept.
