### 2025-06-10 #8076 @archermarx
Got the code installed and working without too much issue. A brief test of CPU strong scaling worked as expected. I think my review is complete and I'm ready to accept.

![Image](https://github.com/user-attachments/assets/915a3956-57c9-41b4-855a-f7b5c153d637)
