### 2025-06-11 #8076 @akashdhruv
@gianlupo I am getting following compilation errors when compiling the code for CPU and GPU versions when I run `make libs && make`:

CPU with openmpi 4.1.1 and gcc 9.4.0:

```
mpifort  -O3 src/initflow.f90.o src/rk.f90.o src/initsolver.f90.o src/chkdiv.f90.o src/initmpi.f90.o src/vof_thinc_qq.f90.o src/workspaces.f90.o src/sanity.f90.o src/fft.f90.o src/chkdt.f90.o src/load.f90.o src/common_mpi.f90.o src/correc.f90.o src/nvtx.f90.o src/fillps.f90.o src/fftw.f90.o src/scal.f90.o src/main.f90.o src/mom.f90.o src/debug.f90.o src/two_fluid.f90.o src/solver.f90.o src/forcing.f90.o src/param.f90.o src/solver_vc.f90.o src/bound.f90.o src/common_cudecomp.f90.o src/initgrid.f90.o src/solver_gpu.f90.o src/acdi.f90.o src/timer.f90.o src/updatep.f90.o src/utils.f90.o src/post.f90.o src/types.f90.o src/output.f90.o  -L./dependencies/2decomp-fft -ldecomp2d -lfftw3  -I./dependencies/2decomp-fft/mod -o ./run/cans
/usr/bin/ld: cannot find -lfftw3
```

GPU:

```
cd ./dependencies/2decomp-fft && make
make[1]: Entering directory '/home/akash/Desktop/CaNS-Fizzy/dependencies/2decomp-fft'
echo "FC = mpifort" > Makefile.settings
echo "FFLAGS = -cpp -std=f2008 -O3 -march=native -J mod -Imod" >> Makefile.settings
echo "OPT = " >> Makefile.settings
echo "DEFS = -DVERSION=\"dc4ce5c\" -DDOUBLE_PREC -cpp" >> Makefile.settings
echo "INC = " >> Makefile.settings
echo "LIBOPT = rcs" >> Makefile.settings
echo "LIBFFT = " >> Makefile.settings
echo "LFLAGS = " >> Makefile.settings
make[1]: Nothing to be done for 'all'.
make[1]: Leaving directory '/home/akash/Desktop/CaNS-Fizzy/dependencies/2decomp-fft'
cd ./dependencies/cuDecomp && mkdir -p build && cd build && cmake .. && make -j
-- The CXX compiler identification is NVHPC 21.3.0
-- The CUDA compiler identification is NVIDIA 11.2.142
-- The Fortran compiler identification is NVHPC 21.3.0
-- Detecting CXX compiler ABI info
-- Check for working CXX compiler: /usr/local/nvhpc-cuda-11.2/Linux_x86_64/21.3/compilers/bin/nvc++ - skipped
-- Detecting CXX compile features
-- Detecting CUDA compiler ABI info
-- Detecting CUDA compiler ABI info - done
-- Check for working CUDA compiler: /usr/local/nvhpc-cuda-11.2/Linux_x86_64/21.3/compilers/bin/nvcc - skipped
-- Detecting CUDA compile features
-- Detecting CUDA compile features - done
-- Detecting Fortran compiler ABI info
-- Detecting Fortran compiler ABI info - done
-- Check for working Fortran compiler: /usr/local/nvhpc-cuda-11.2/Linux_x86_64/21.3/compilers/bin/nvfortran - skipped
-- Found MPI_CXX: /usr/local/nvhpc-cuda-11.2/Linux_x86_64/21.3/comm_libs/openmpi/openmpi-3.1.5/lib/libmpi_cxx.so (found version "3.1")
-- Could NOT find MPI_Fortran (missing: MPI_Fortran_WORKS)
CMake Error at /home/akash/.local/lib/python3.8/site-packages/cmake/data/share/cmake-3.26/Modules/FindPackageHandleStandardArgs.cmake:230 (message):
  Could NOT find MPI (missing: MPI_Fortran_FOUND) (found version "3.1")
Call Stack (most recent call first):
  /home/akash/.local/lib/python3.8/site-packages/cmake/data/share/cmake-3.26/Modules/FindPackageHandleStandardArgs.cmake:600 (_FPHSA_FAILURE_MESSAGE)
  /home/akash/.local/lib/python3.8/site-packages/cmake/data/share/cmake-3.26/Modules/FindMPI.cmake:1837 (find_package_handle_standard_args)
  CMakeLists.txt:54 (find_package)

-- Configuring incomplete, errors occurred!
make: *** [dependencies/external.mk:7: libs] Error 1
```

Maybe I am missing something.
