### 2025-06-15 #8076 @p-costa
Hi @akashdhruv,

Let me chime in here as co-author of the manuscript:

1. For the CPU build, it seems that the linker cannot find the FFTW3 library. This is one of the required dependencies - do you have it installed? On Ubuntu, you can check the steps in the CI to be able to compile the code on CPU: https://github.com/CaNS-World/CaNS-Fizzy/blob/main/.github/workflows/compile.yml#L15

2. For the GPU build, it seems that you are unable to build the [`cuDecomp`](https://github.com/NVIDIA/cuDecomp) library, a dependency for the GPU build, with `cmake`. I wonder if you have an MPI library built along with the NVHPC installation.

    - On Ubuntu, you can locally install NVHPC with everything else you need using this script: [link](https://github.com/p-costa/CaNS-test-CI/blob/refactored-ci/.github/actions/build/scripts/install-NVIDIA.sh).

    - Then, to be sure all the environment variables are set, you can set up the environment like this, which is in line with what is stated in the NVHPC docs: [link](https://github.com/p-costa/CaNS-test-CI/blob/refactored-ci/.github/actions/build/scripts/setenv-NVIDIA.sh).
