### 2025-06-30 #8076 @gianlupo
Hi @akashdhruv,

any update for the software compilation?
