### 2025-07-10 #8076 @akashdhruv
@p-costa @gianlupo

- The CPU version compiles after installing FFTW3. Thank you for pointing that out.
- For GPU version, I was already using NVHPC compiler but the MPI Fortran compiler is not being detected by the Cmake system for some reason even though it's present in the path. I will work on this a bit more, but I don't find success I will approve the paper because I am fairly satisfied with the CPU version of the code.
