### 2025-07-23 #8076 @gianlupo
> I’m curious why the manuscript does not include a comparison with Flash-X (https://flash-x.org/), which is an AMR-based code designed for incompressible multiphase flow simulations. Flash-X uses AMReX as an option for grid backend and multigrid Poisson solver. Including such a comparison, either in performance, capabilities, or intended use cases, could provide valuable context for readers. For reference, here are some relevant materials:
>
>     * https://www.sciencedirect.com/science/article/abs/pii/S0021999124003711
>
>     * https://www.sciencedirect.com/science/article/pii/S2352711022001030
>
>     * https://www.youtube.com/watch?v=pWio2Jy0PUA
>
>
> This suggestion is offered in the interest of completeness and clarity, not self-promotion.

Dear @akashdhruv,
thank you for your review. Regarding your suggestion about Flash-X, thank you for bringing to our attention the code's capabilities and where its applicability overlaps with CaNS-Fizzy. We will certainly consider it as point of comparison in the future when working with CaNS-Fizzy on more physically-oriented publications. For the scope of the present work however, our intent is to briefly place CaNS-Fizzy within the general context of multiphase flow simulation tools currently available, for which purpose we think that referring to the two most widely used open source softwares (openfoam and basilisk) is sufficient.
