+++ paper/paper.bib
@@ -121,6 +121,7 @@ author = {F. Municchi and C.N. Markides and O.K. Matar and M. Magnini}
 @article{flow36,
 title = {FLOW36: A spectral solver for phase-field based multiphase turbulence simulations on heterogeneous computing architectures},
 journal = {Computer Physics Communications},
+volume = {313},
 pages = {109640},
 year = {2025},
 doi = {10.1016/j.cpc.2025.109640},
