### 2025-07-31 #8076 @gianlupo
> [@gianlupo](https://github.com/gianlupo) [@p-costa](https://github.com/p-costa) Here are some errors in your `paper.md` that needs to be fixed:
>
>     1. Software name at line 35 in the rendered PDF: **boilingfoam** -> **boilingFoam**
>
>     2. Wrong usage for multiple citation: please change `[@gerris,@basilisk]` to `[@gerris; @basilisk]`. If you check line 35 in the PDF, you can see that currently these two citations are not rendered correctly.
>
>     3. Incomplete reference/bibtex record: the reference for `flow36` is incomplete. The `volume` is missing.

Dear @piyueh,

thank you for pointing out these mistakes! I pushed a new commit that fixes them.
