### 2025-08-04 #8076 @piyueh
@gianlupo @p-costa We are now proceeding to the final final step. We need you to make a tagged release of the software, and deposit a copy of the repository with a data-archiving service such as [Zenodo](https://zenodo.org/) or [figshare](https://figshare.com/). Once you have done with this step, please post the version number and the DOI of either Zenodo or figshare here. Please let me know if there is any question. Thanks!
