### 2025-08-13 #8076 @piyueh
@p-costa The title of the Zenodo archive should match the JOSS paper (i.e., CaNS-Fizzy: A GPU-accelerated finite difference solver for turbulent two-phase flows)
