### 2025-08-14 #8076 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
